Concurrency-related code refactored and moved to Async.kt
* Threads are now created and maintained by a cached thread pool
* concurrency-related functions and helpers moved from Component.kt, lib.kt and FX.kt to Async.kt
* `runAsync` and `ui` methods can now be called from anywhere
* `runAsync` kept in UIComponent because of scope usage, so if called from within UIComponent it's behavior will remain unaltered.
* `runAsyncWithProgress` supports latch
* `runAsyncWithOverlay` some minor changes to docs
