+++ src/main/java/tornadofx/FX.kt
@@ -2,12 +2,12 @@
 
 package tornadofx
 
-import com.sun.javafx.tk.Toolkit
 import javafx.application.Application
 import javafx.application.Platform
-import javafx.beans.property.*
-import javafx.beans.value.ChangeListener
-import javafx.beans.value.ObservableValue
+import javafx.beans.property.ListProperty
+import javafx.beans.property.SimpleBooleanProperty
+import javafx.beans.property.SimpleObjectProperty
+import javafx.beans.property.SimpleStringProperty
 import javafx.collections.FXCollections
 import javafx.collections.ListChangeListener
 import javafx.collections.ObservableList
@@ -26,7 +26,6 @@ import javafx.scene.layout.HBox
 import javafx.scene.layout.Pane
 import javafx.scene.layout.VBox
 import javafx.stage.Stage
-import javafx.util.Duration
 import tornadofx.FX.Companion.inheritParamHolder
 import tornadofx.FX.Companion.inheritScopeHolder
 import tornadofx.FX.Companion.stylesheets
@@ -626,86 +625,4 @@ private fun Parent.getChildrenReflectively(): MutableList<Node>? {
         return getter.invoke(this) as MutableList<Node>
     }
     return null
-}
-
-
-/**
- * Run the specified Runnable on the JavaFX Application Thread at some
- * unspecified time in the future.
- */
-fun runLater(op: () -> Unit) = Platform.runLater(op)
-
-/**
- * Run the specified Runnable on the JavaFX Application Thread after a
- * specified delay.
- *
- * runLater(10.seconds) {
- *     // Do something on the application thread
- * }
- *
- * This function returns a TimerTask which includes a runningProperty as well as the owning timer.
- * You can cancel the task before the time is up to abort the execution.
- */
-fun runLater(delay: Duration, op: () -> Unit): FXTimerTask {
-    val timer = Timer(true)
-    val task = FXTimerTask(op, timer)
-    timer.schedule(task, delay.toMillis().toLong())
-    return task
-}
-
-class FXTimerTask(val op: () -> Unit, val timer: Timer) : TimerTask() {
-    private val internalRunning = ReadOnlyBooleanWrapper(false)
-    val runningProperty: ReadOnlyBooleanProperty get() = internalRunning.readOnlyProperty
-    val running: Boolean get() = runningProperty.value
-
-    private val internalCompleted = ReadOnlyBooleanWrapper(false)
-    val completedProperty: ReadOnlyBooleanProperty get() = internalCompleted.readOnlyProperty
-    val completed: Boolean get() = completedProperty.value
-
-    override fun run() {
-        internalRunning.value = true
-        Platform.runLater {
-            try {
-                op()
-            } finally {
-                internalRunning.value = false
-                internalCompleted.value = true
-            }
-        }
-    }
-}
-
-/**
- * Wait on the UI thread until a certain value is available on this observable.
- *
- * This method does not block the UI thread even though it halts further execution until the condition is met.
- */
-fun <T> ObservableValue<T>.awaitUntil(condition: (T) -> Boolean) {
-    if (!Toolkit.getToolkit().canStartNestedEventLoop()) {
-        throw IllegalStateException("awaitUntil is not allowed during animation or layout processing")
-    }
-
-    val changeListener = object : ChangeListener<T> {
-        override fun changed(observable: ObservableValue<out T>?, oldValue: T, newValue: T) {
-            if (condition(value)) {
-                runLater {
-                    Toolkit.getToolkit().exitNestedEventLoop(this@awaitUntil, null)
-                    removeListener(this)
-                }
-            }
-        }
-    }
-
-    changeListener.changed(this, value, value)
-    addListener(changeListener)
-    Toolkit.getToolkit().enterNestedEventLoop(this)
-}
-
-/**
- * Wait on the UI thread until this observable value is true.
- *
- * This method does not block the UI thread even though it halts further execution until the condition is met.
- */
-fun ObservableValue<Boolean>.awaitUntil() {
-    this.awaitUntil { it }
 }
