+++ src/main/java/tornadofx/Lib.kt
@@ -1,13 +1,11 @@
 package tornadofx
 
-import javafx.application.Platform
 import javafx.beans.InvalidationListener
 import javafx.beans.property.*
 import javafx.beans.value.*
 import javafx.collections.*
 import javafx.collections.transformation.FilteredList
 import javafx.collections.transformation.SortedList
-import javafx.concurrent.Task
 import javafx.geometry.Insets
 import javafx.scene.control.ListView
 import javafx.scene.control.TableView
@@ -207,76 +205,6 @@ class SortedFilteredList<T>(
 fun <T> List<T>.observable(): ObservableList<T> = FXCollections.observableList(this)
 fun <T> Set<T>.observable(): ObservableSet<T> = FXCollections.observableSet(this)
 fun <K, V> Map<K, V>.observable(): ObservableMap<K, V> = FXCollections.observableMap(this)
-
-class FXTask<T>(val status: TaskStatus? = null, val func: FXTask<*>.() -> T) : Task<T>() {
-    val completedProperty: ReadOnlyBooleanProperty = SimpleBooleanProperty(false)
-    val completed by completedProperty
-
-    override fun call() = func(this)
-
-    init {
-        status?.item = this
-    }
-
-    override fun succeeded() {
-        (completedProperty as BooleanProperty).value = true
-    }
-
-    override fun failed() {
-        (completedProperty as BooleanProperty).value = true
-    }
-
-    override fun cancelled() {
-        (completedProperty as BooleanProperty).value = true
-    }
-
-    override public fun updateProgress(workDone: Long, max: Long) {
-        super.updateProgress(workDone, max)
-    }
-
-    override public fun updateProgress(workDone: Double, max: Double) {
-        super.updateProgress(workDone, max)
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    fun value(v: Any) {
-        super.updateValue(v as T)
-    }
-
-    override public fun updateTitle(t: String?) {
-        super.updateTitle(t)
-    }
-
-    override public fun updateMessage(m: String?) {
-        super.updateMessage(m)
-    }
-
-}
-
-open class TaskStatus : ItemViewModel<FXTask<*>>() {
-    val running: ReadOnlyBooleanProperty = bind { SimpleBooleanProperty().apply { if (item != null) bind(item.runningProperty()) } }
-    val completed: ReadOnlyBooleanProperty = bind { SimpleBooleanProperty().apply { if (item != null) bind(item.completedProperty) } }
-    val message: ReadOnlyStringProperty = bind { SimpleStringProperty().apply { if (item != null) bind(item.messageProperty()) } }
-    val title: ReadOnlyStringProperty = bind { SimpleStringProperty().apply { if (item != null) bind(item.titleProperty()) } }
-    val progress: ReadOnlyDoubleProperty = bind { SimpleDoubleProperty().apply { if (item != null) bind(item.progressProperty()) } }
-}
-
-fun <T> task(taskStatus: TaskStatus? = null, func: FXTask<*>.() -> T): Task<T> = FXTask(taskStatus, func = func).apply {
-    setOnFailed({ Thread.getDefaultUncaughtExceptionHandler().uncaughtException(Thread.currentThread(), exception) })
-    Thread(this).start()
-}
-
-infix fun <T> Task<T>.success(func: (T) -> Unit) = apply {
-    Platform.runLater {
-        setOnSucceeded { func(value) }
-    }
-}
-
-infix fun <T> Task<T>.fail(func: (Throwable) -> Unit) = apply {
-    Platform.runLater {
-        setOnFailed { func(exception) }
-    }
-}
 
 fun Clipboard.setContent(op: ClipboardContent.() -> Unit) {
     val content = ClipboardContent()
